test: add unified test_main
