+++ firmware/platformio.ini
@@ -166,7 +166,6 @@ build_src_filter =
     +<**/*.cpp>
     -<**/test_*.cpp>
     +<include/**>
-    +<../test/TestHelpers.cpp>
 lib_deps =
     ${env:teensy40_base.lib_deps}
     unity           ; keep <unity.h> on the path so pio run -e teensy40_unity doesn't bomb
