+++ firmware/test/test_display_manager.cpp
@@ -7,13 +7,3 @@ void test_update_interval_round_trip() {
     TEST_ASSERT_EQUAL_UINT32(1234, dm.getUpdateInterval());
 }
 
-void setUp(void) {}
-void tearDown(void) {}
-
-void setup() {
-    UNITY_BEGIN();
-    RUN_TEST(test_update_interval_round_trip);
-    UNITY_END();
-}
-
-void loop() {}
