+++ firmware/test/test_potentiometer_manager.cpp
@@ -9,13 +9,3 @@ void test_channel_and_cc() {
     TEST_ASSERT_EQUAL_UINT8(74, pm.getCCNumber(0));
 }
 
-void setUp(void) {}
-void tearDown(void) {}
-
-void setup() {
-    UNITY_BEGIN();
-    RUN_TEST(test_channel_and_cc);
-    UNITY_END();
-}
-
-void loop() {}
